requirements.txt
streamlit
python-dotenv
langchain
openai
tabulate
chromadb
tiktoken
flask
Flask-SQLAlchemy
mysqlclient
duckduckgo-search